Anmend merge

+++ modules/core/PrimSphere.cpp
@@ -82,7 +82,7 @@ namespace rt {
 		// Vec3f hitPoint = ray.hitPoint() - m_origin;	// Hitpoint in WCS
 		// float r = m_radius;
 		Vec3f hitPoint = wcs2ocs(ray.hitPoint());		// Hitpoint in OCS
-		float r = norm(hitPoint);						// sqrt(x^2 + y^2 + z^2)
+		float r = static_cast<float>(norm(hitPoint));	// sqrt(x^2 + y^2 + z^2)
 		float phi = atan2f(hitPoint[2], hitPoint[0]);	// arctg(z / x)
 		float theta = acosf(MIN(r, hitPoint[1]) / r);	// arccos(y / r)
 		return Vec2f(-0.5f * phi / Pif, theta / Pif);
